+++ common/src/main/java/stirling/software/common/configuration/InstallationPathConfig.java
