+++ common/src/main/java/stirling/software/common/model/ApplicationProperties.java
