+++ common/src/main/java/stirling/software/common/model/InputStreamTemplateResource.java
