+++ common/src/main/java/stirling/software/common/model/enumeration/UsernameAttribute.java
