+++ common/src/main/java/stirling/software/common/model/exception/UnsupportedProviderException.java
