+++ src/main/java/stirling/software/SPDF/config/security/mail/MailConfig.java
@@ -11,7 +11,7 @@ import org.springframework.mail.javamail.JavaMailSenderImpl;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import stirling.software.SPDF.model.ApplicationProperties;
+import stirling.software.common.model.ApplicationProperties;
 
 /**
  * This configuration class provides the JavaMailSender bean, which is used to send emails. It reads
