+++ src/main/resources/application.properties
@@ -21,7 +21,7 @@ spring.servlet.multipart.max-file-size=2000MB
 spring.servlet.multipart.max-request-size=2000MB
 server.servlet.session.tracking-modes=cookie
 server.servlet.context-path=${SYSTEM_ROOTURIPATH:/}
-spring.devtools.restart.enabled=true
+spring.devtools.restart.enabled=false
 spring.devtools.livereload.enabled=true
 spring.devtools.restart.exclude=stirling.software.SPDF.config.security/**
 spring.thymeleaf.encoding=UTF-8
